requirements.txt
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
flask-restplus==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
SQLAlchemy==1.1.15
sqlalchemy-migrate==0.11.0
pytest==3.3.0
flipflop==1.0
guess-language==0.2
coverage==4.4.2
celery==4.1.0
